Refactoring imports, cleanup
